backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
openai>=1.0.0
requests
python-multipart
streamlit
faiss-cpu
PyMuPDF
numpy